refactor: extract shared jwt decode utility from duplicated code
